requirements-api.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
sentence-transformers>=2.6.0
huggingface_hub
qdrant-client==1.7.0